+++ pytest.ini
@@ -33,3 +33,4 @@ filterwarnings=
 	ignore:pkg_resources is deprecated as an API:DeprecationWarning
 	ignore:module 'sre_constants' is deprecated:DeprecationWarning:pkg_resources._vendor.pyparsing
 	ignore:.+ is shallow and may cause errors:UserWarning:setuptools_scm.git
+pythonpath=test_support
