+++ setup.cfg
@@ -43,7 +43,7 @@ where = src
 [options.extras_require]
 testing =
 	# upstream
-	pytest >= 6
+	pytest >= 7
 	pytest-console-scripts >= 1.2
 	pytest-cov
 	pytest-enabler >= 2.2; \
